requirements.txt
colorama
dateparser
discord.py
facebook-sdk
praw
pymongo
PyTumblr
pytz
pyNaCL
python-twitter
requests
ruamel.yaml
tweepy
twitter
wiktionaryparser
